fix: reflect team feedback
